Fixed MemoryOps detection of bitwise operations

+++ include/Kor/Core/Unix/UnixMemory.h
@@ -33,7 +33,7 @@ struct SUnixMemoryOps
 	KOR_FORCEINLINE static void* Move(void* dest, const void* src, uint64 size) noexcept { return memmove(dest, src, size); }
 	
 	// Fills block of memory with specific value
-	KOR_FORCEINLINE static void* Fill(void* dest, int32 val, uint64 size) noexcept { return memset(dest, val, size); }
+	KOR_FORCEINLINE static void* Fill(void* dest, uint8 val, uint64 size) noexcept { return memset(dest, val, size); }
 
 	// Fills block of memory with zero value
 	KOR_FORCEINLINE static void* Zero(void* dest, uint64 size) noexcept { return memset(dest, 0, size); }
